+++ src/CoreShop/Bundle/CoreBundle/Pimcore/Repository/WishlistRepository.php
@@ -18,6 +18,7 @@ declare(strict_types=1);
 
 namespace CoreShop\Bundle\CoreBundle\Pimcore\Repository;
 
+use Carbon\Carbon;
 use CoreShop\Component\Core\Wishlist\Repository\WishlistRepositoryInterface;
 use CoreShop\Component\Customer\Model\CustomerInterface;
 use CoreShop\Component\Store\Model\StoreInterface;
@@ -42,5 +43,39 @@ class WishlistRepository extends \CoreShop\Bundle\WishlistBundle\Pimcore\Reposit
         }
 
         return null;
+    }
+
+    public function findExpiredWishlists(int $days, bool $anonymous, bool $customer): array
+    {
+        $daysTimestamp = Carbon::now();
+        $daysTimestamp->subDays($days);
+
+        $conditions = ['o_modificationDate < ?'];
+        $params = [$daysTimestamp->getTimestamp()];
+        $groupCondition = [];
+
+        if (true === $anonymous) {
+            $groupCondition[] = 'customer__id IS NULL';
+        }
+
+        if (true === $customer) {
+            $groupCondition[] = 'customer__id IS NOT NULL';
+        }
+
+        $bind = ' AND ';
+        $groupBind = ' OR ';
+
+        $sql = implode($bind, $conditions);
+        $sql .= ' AND (' . implode($groupBind, $groupCondition) . ') ';
+
+        $list = $this->getList();
+        $list->setCondition($sql, $params);
+
+        /**
+         * @var WishlistInterface[] $result
+         */
+        $result = $list->getObjects();
+
+        return $result;
     }
 }
