+++ src/CoreShop/Bundle/WishlistBundle/Pimcore/Repository/WishlistRepository.php
@@ -18,7 +18,6 @@ declare(strict_types=1);
 
 namespace CoreShop\Bundle\WishlistBundle\Pimcore\Repository;
 
-use Carbon\Carbon;
 use CoreShop\Bundle\ResourceBundle\Pimcore\PimcoreRepository;
 use CoreShop\Component\StorageList\Model\StorageListInterface;
 use CoreShop\Component\Wishlist\Model\WishlistInterface;
@@ -44,40 +43,5 @@ class WishlistRepository extends PimcoreRepository implements WishlistRepository
         }
 
         return null;
-    }
-
-    public function findExpiredWishlists(int $days, bool $anonymous, bool $customer): array
-    {
-        $daysTimestamp = Carbon::now();
-        $daysTimestamp->subDays($days);
-        
-        $conditions = ['o_modificationDate < ?'];
-        $params = [$daysTimestamp->getTimestamp()];
-        $groupCondition = [];
-
-        if (true === $anonymous) {
-            $groupCondition[] = 'customer__id IS NULL';
-        }
-
-        if (true === $customer) {
-            $groupCondition[] = 'customer__id IS NOT NULL';
-        }
-
-        $bind = ' AND ';
-        $groupBind = ' OR ';
-
-        $sql = implode($bind, $conditions);
-        $sql .= ' AND (' . implode($groupBind, $groupCondition) . ') ';
-
-        $list = $this->getList();
-        $list->setCondition($sql, $params);
-
-        /**
-         * @var WishlistInterface[] $result
-         */
-        $result = $list->getObjects();
-
-        return $result;
-
     }
 }
