+++ src/CoreShop/Bundle/WishlistBundle/Resources/config/services.yml
@@ -32,6 +32,7 @@ services:
         tags:
             - { name: coreshop.context.wishlist, priority: -999 }
 
+
     CoreShop\Bundle\WishlistBundle\Form\Type\WishlistType:
         arguments:
             - '%coreshop.model.wishlist.class%'
